Backend/requirements.txt
Django==2.0
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.10
pytz==2017.3
